test: complete test for table "currencies" and "transaction_type"

+++ database/migrations/20250824080218_create_table_categories_and_transaction_types.up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE transaction_types (
-    id      integer     PRIMARY KEY,
+    id      smallint    PRIMARY KEY,
     name    text        NOT NULL
 );
